% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtureMetaC.R
\name{MixtureMetaC}
\alias{MixtureMetaC}
\title{Mixture population model approach for meta analysis of continuous outcome}
\usage{
MixtureMetaC(DSN = NULL, WEIGHT = NULL, ALPHA = 0.05, SE = FALSE, DEMOG = NULL)
}
\arguments{
\item{DSN}{Input data frame}

\item{WEIGHT}{Weights used to combine studies: '"size"', '"equal"' or a numeric vector}

\item{ALPHA}{Two-sided significance level}

\item{SE}{'TRUE' for standard error included in input data frame instead of standard deviation and 'FALSE' (the default) otherwise}

\item{DEMOG}{A character vector for baseline characteristics included in input data frame}
}
\value{
A data frame
}
\description{
Mixture population model approach for meta analysis of continuous outcome
}
\examples{
MixtureMetaC(DSN = ELDL, WEIGHT = "size", DEMOG = c("Age", "Male", "LDL"))
MixtureMetaC(DSN = ELDL, WEIGHT = "equal")
MixtureMetaC(DSN = ELDL, WEIGHT = rep(1, 16))
}
