% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtureMetaB.R
\name{MixtureMetaB}
\alias{MixtureMetaB}
\title{Mixture population model approach for meta analysis of binary outcome}
\usage{
MixtureMetaB(
  DSN = NULL,
  WEIGHT = NULL,
  ALPHA = 0.05,
  NBOOT = 2000,
  DEMOG = NULL
)
}
\arguments{
\item{DSN}{Input data frame}

\item{WEIGHT}{Weights used to combine studies: '"size"', '"equal"' or a numeric vector}

\item{ALPHA}{Two-sided significance level}

\item{NBOOT}{Number of iterations. When default (NBOOT = 2000), 2,000 sets of realizations are generated for calculating confidence intervals of risk ratio estimated by the mixture population model approach.}

\item{DEMOG}{A character vector for baseline characteristics included in input data frame}
}
\value{
A data frame
}
\description{
Mixture population model approach for meta analysis of binary outcome
}
\examples{
MixtureMetaB(DSN = EHFE, WEIGHT = "size", DEMOG = c("AGE", "MALE", "BMI", "DD", "FU"))
MixtureMetaB(DSN = EHFE, WEIGHT = "equal")
MixtureMetaB(DSN = EHFE, WEIGHT = rep(1, 6))
}
